Progress Kelola 5 Surat

+++ backend/app/Http/Controllers/SuratKeteranganKerjaController.php
@@ -6,9 +6,12 @@ use App\Http\Resources\SuratKeteranganKerjaCollection as ThisCollection;
 use App\Models\ReferenceNumberSetting;
 use App\Models\SuratKeteranganKerja as Letter;
 use Illuminate\Http\Request;
+use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
+use NcJoes\OfficeConverter\OfficeConverter;
 
 class SuratKeteranganKerjaController extends Controller
 {
@@ -32,6 +35,7 @@ class SuratKeteranganKerjaController extends Controller
             'employee.position' => 'required|string',
             'signer.nip' => 'required|exists:employees,nip',
             'signer.position' => 'required|string',
+            'signature_type' => 'required|in:manual,qrcode,digital',
         ]);
 
         if ($validate->fails()) {
@@ -48,6 +52,7 @@ class SuratKeteranganKerjaController extends Controller
         $letter->position = $request->employee['position'];
         $letter->signer_nip = $request->signer['nip'];
         $letter->signer_position = $request->signer['position'];
+        $letter->signature_type = $request->signature_type;
         $letter->created_by = auth()->user()->nip;
         $letter->save();
 
@@ -74,33 +79,44 @@ class SuratKeteranganKerjaController extends Controller
         ], 200);
     }
 
-    public function download(string $id) {
+    public function download(string $id)
+    {
         $letter = Letter::find($id);
         if (!$letter) {
             return response()->json([
                 'message' => "Surat keterangan kerja tidak ditemukan !"
             ], 404);
         }
+
+        if ($letter->signed_file != null) {
+            $fileNameServerPdf = 'app\\signed_files\\' . $letter->signed_file;
+            return response()->download(storage_path($fileNameServerPdf), $letter->signed_file);
+        }
+
+        $filename = $letter->id . '.pdf';
+        $fileNameServerPdf = 'app/tmp/surat_keterangan_kerja/' . $filename;
+
+        if ($letter->tmp_file && file_exists(storage_path($letter->tmp_file))) {
+            return response()->download(storage_path($letter->tmp_file), $filename);
+        }
+
         $templateProcessor = $letter->generate_docx();
-        $fileNameServerDocx = 'app/tmp/' . Str::replace("/", "-", $letter->get_reference_number()) . '.docx';
+        $fileNameServerDocx = "app\\tmp\\surat_keterangan_kerja\\" . $letter->id . '.docx';
         $templateProcessor->saveAs(storage_path($fileNameServerDocx));
 
-        $domPdfPath = base_path('vendor/dompdf/dompdf');
-        \PhpOffice\PhpWord\Settings::setPdfRendererPath($domPdfPath);
-        \PhpOffice\PhpWord\Settings::setPdfRendererName('DomPDF');
-         
-        //Load word file
-        $Content = \PhpOffice\PhpWord\IOFactory::load(storage_path($fileNameServerDocx)); 
- 
-        //Save it into PDF
-        $fileNameServerPdf = 'app/tmp/' . Str::replace("/", "-", $letter->get_reference_number()) . '.pdf';
-        $PDFWriter = \PhpOffice\PhpWord\IOFactory::createWriter($Content,'PDF');
-        $PDFWriter->save(storage_path($fileNameServerPdf));
+        $response = Http::post(env('APP_DOCX_CONVERTER_URL') . '/convert', ['file_path' => storage_path($fileNameServerDocx)]);
+        if ($response->failed()) {
+            return response()->json([
+                'errors' => "Something errors"
+            ], 500);
+        }
 
-        unlink(storage_path($fileNameServerDocx));
-
-        $filename = Str::replace("/", "-", $letter->get_reference_number()) . '.pdf';
-        return response()->download(storage_path($fileNameServerPdf), $filename)->deleteFileAfterSend(true);
+        if ($response->successful() && file_exists(storage_path($fileNameServerPdf))) {
+            unlink(storage_path($fileNameServerDocx));
+            $letter->tmp_file = $fileNameServerPdf;
+            $letter->save();
+            return response()->download(storage_path($fileNameServerPdf), $filename);
+        }
     }
 
     /**
@@ -111,6 +127,47 @@ class SuratKeteranganKerjaController extends Controller
         //
     }
 
+    public function upload_signed_file(Request $request)
+    {
+        $validate = Validator::make($request->all(), [
+            'signed_file' => 'required|file|mimes:pdf|max:10000',
+        ]);
+
+        if ($validate->fails()) {
+            $response = [
+                'errors' => $validate->errors(),
+                'message' => "Validasi form gagal !"
+            ];
+            return response()->json($response, 422);
+        }
+
+        $letter = Letter::find($request->id);
+        if (!$letter) {
+            return response()->json([
+                'message' => "Surat keterangan kerja tidak ditemukan !"
+            ], 404);
+        }
+
+        if (!$letter->can_upload_verified_file()) {
+            return response()->json([
+                'message' => "Surat dengan jenis tanda tangan selain manual dan digital tidak dapat ditandatangani dengan !"
+            ], 403);
+        }
+
+        $file = $request->file('signed_file');
+        $fileName = Str::replace("/", "-", $letter->get_reference_number()) . '.' . $file->getClientOriginalExtension();
+        $fileNameServer = 'signed_files/' . $fileName;
+        Storage::put($fileNameServer, file_get_contents($file));
+
+        $letter->signed_file = $fileName;
+        $letter->save();
+
+        $response = [
+            'message' => "Berhasil mengupload file surat keterangan kerja !"
+        ];
+        return response()->json($response, 200);
+    }
+
     /**
      * Remove the specified resource from storage.
      */
@@ -119,7 +176,8 @@ class SuratKeteranganKerjaController extends Controller
         //
     }
 
-    public function give_reference_number(Request $request) {
+    public function give_reference_number(Request $request)
+    {
         $letter = Letter::find($request->id);
         if (!$letter) {
             return response()->json([
@@ -128,13 +186,15 @@ class SuratKeteranganKerjaController extends Controller
         }
 
         //get the latest reference_number
-        $latest_number = Letter::where('reference_number', "IS NOT NULL")->count();
+        $latest_number = Letter::where('reference_number', '!=', 'NULL')->whereMonth('created_at', Carbon::now()->month)->whereYear('created_at', Carbon::now()->year)->count();
         $letter->reference_number = ReferenceNumberSetting::get_and_parse_reference_number_with_date(Letter::NAME, $latest_number + 1, $letter->created_at);
+        $letter->tmp_file = null;
         $letter->save();
 
         $response = [
             'message' => "Berhasil memberikan nomor surat !"
         ];
+
         return response()->json($response, 200);
     }
 }
